Fix mocked Version_Dialog test failures using conftest.py\n\nCI pipelines failed because tests were loading and interacting directly with PyQt widgets and mocked classes asynchronously. Moved massive dependency mocking routines to `tests/conftest.py` with `@pytest.fixture(autouse=True, scope="session")` to cleanly inject mocked `MagicMock()` systems globally. Removed dirty patch assignments from `tests/test_battle_handler.py`.
